+++ example/lib/screens/home_screen.dart
@@ -96,7 +96,6 @@ class _FormUsageScreenState extends PlexState<FormUsageScreen> {
             hideLoading();
           },
         );
-        print(myUser.toString());
       },
     );
   }