+++ lib/plex_annotations/plex_annotation_builders/plex_annotation_builders.dart
@@ -1,10 +1,10 @@
-import 'package:build/build.dart';
-import 'package:plex/plex_annotations/plex_generators/plex_model_generators.dart';
-import 'package:source_gen/source_gen.dart';
-
-Builder generateModel(BuilderOptions options) {
-  return PartBuilder(
-    [PlexModelGenerator()],
-    '.plex.dart',
-  );
-}
+// import 'package:build/build.dart';
+// import 'package:plex/plex_annotations/plex_generators/plex_model_generators.dart';
+// import 'package:source_gen/source_gen.dart';
+//
+// Builder generateModel(BuilderOptions options) {
+//   return PartBuilder(
+//     [PlexModelGenerator()],
+//     '.plex.dart',
+//   );
+// }