+++ lib/plex_assets.dart
@@ -1,3 +1,5 @@
 const String pdf = "packages/plex/assets/pdf.png";
 const String excel = "packages/plex/assets/excel.png";
 const String groupData = "packages/plex/assets/group_data.png";
+const String loadingV1 = "packages/plex/assets/loading/loading_1.json";
+const String loadingV2 = "packages/plex/assets/loading/loading_2.json";