+++ lib/plex_widgets/loading/plex_loading_enum.dart
@@ -1,0 +1,4 @@
+enum PlexLoadingEnum {
+  version1,
+  version2,
+}