+++ macos/Flutter/GeneratedPluginRegistrant.swift
@@ -9,7 +9,7 @@ import connectivity_plus
 import file_saver
 import path_provider_foundation
 import shared_preferences_foundation
-import sqflite
+import sqflite_darwin
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
   ConnectivityPlusPlugin.register(with: registry.registrar(forPlugin: "ConnectivityPlusPlugin"))